data_service/requirements.txt
paho-mqtt==1.6.1
clickhouse-connect==0.6.23
python-dateutil==2.8.2
